Remove Azure tools.py
